modules/modeling/basic/z-score/requirements.txt
pyDataCanvas >= 0.4.0,<0.5.0
numpy
scikit-learn
